Separating conversion logic for models and props (jsInterop)
